Build with PostgreSQL 10
Move over to the standard PostgreSQL build system for building libpq since
the old method of using nmake has been removed.  Requires modifing a file
so that the lib is built instead of the dll.  Libpgcommon and libpgport is
also required to be built and linked to the resulting library.