on_success to deploy built files if on psycopg2 account
To assisting debugging when not on the main psycopg2 account, add a check
for the REMOTE_KEY environment variable, creating a dummy value for the
key if it doesn't exist.  Also, check to make sure we are on the psycopg2
account before attempting to rsync the newly created file over.

If on a different account than psycopg2, REMOTE_KEY will not be defined
because the decryption of the key will fail.