Build separate binary packages on Linux/OSX
The plan is to release wheel packages as psycopg2-binary to allow
installing them only if desired. For the moment we keep distributing
wheels for the psycopg2 package too, to allow a transition period, but
the resulting package will issue a warning on import. From version 2.8
binary package will only be distributed separately (at least on
Linux/OSX).

We will release windows binary package too: TODO.